Start implementation of syscall handler
Started the implementation of the syscall handler to read the syscall number from the stack and determine which function to call. Also started implementation of reading the needed arguments to pass to each system call

TODO:
Finish the implementation of the logic needed to read needed arguments from the user stack

Test to ensure reading correct information from the stack.